Refactor the ClusterApps component to TypeScript
